requirements.txt
Flask==3.0.1
openai==1.9.0
python-dotenv==1.0.0
PyMuPDF==1.23.16
datasets==2.16.1
transformers==4.36.2
numpy==1.26.3
pandas==2.2.0
pdf2image==1.17.0
opencv-python==4.9.0.80
pytesseract==0.3.10
pyphen==0.14.0
torch==2.1.2